[refactor] folder reestructuring
