+++ rust/src/emergency/handlers.rs
@@ -2,13 +2,15 @@ use super::{send_emergency_notifications, update_user_state};
 use amiquip::Connection as RabbitConnection;
 
 use crate::{
-    cors::options,
     db::{get_connection, get_pool, unlock_channel},
-    catchers::catchers,
+    messaging::get_rabbitmq_uri,
+    middleware::{catchers::catchers, cors::options},
     model::{
-        APIJsonResponse, APIResponse, APIResult, AuthInfo, Message, Storage, UpdateState, UserState,
+        auth::AuthInfo,
+        emergency::{UpdateState, UserState},
+        responses::{APIJsonResponse, APIResponse},
+        APIResult, Message, Storage,
     },
-    publish_websocket_messages::get_rabbitmq_uri,
 };
 use log::error;
 use rocket::{Rocket, State};
