+++ rust/src/http_gateway/handlers.rs
@@ -4,20 +4,23 @@ use amiquip::Connection;
 use rocket::{Rocket, State};
 use rocket_contrib::json::Json;
 
-use crate::cors;
+use crate::middleware::{catchers::catchers, cors::self};
 use crate::db::devices::{get_device_by_id, update_device_settings};
 use crate::db::get_pool;
 use crate::db::telemetry::{
     close_command, force_refresh_telemetry_internal, get_connections, get_follower_keys,
     get_user_state, store_telemetry, username_has_follower,
 };
-use crate::catchers::catchers;
 use crate::model::{
-    APIJsonResponse, APIResponse, AccessType, AuthInfo, CommandResponse, CommandState,
-    DeviceUpdateRequest, DeviceUpdateResponse, FollowerKey, Storage, TelemetryRequest,
-    TelemetryResponse, UserState,
+    responses::{APIJsonResponse, APIResponse, TelemetryResponse, CommandResponse, DeviceUpdateResponse},
+    telemetry::{CommandState, FollowerKey},
+    emergency::AccessType,
+    emergency::UserState,
+    auth::AuthInfo,
+    requests::{DeviceUpdateRequest, TelemetryRequest},
+    Storage,
 };
-use crate::publish_websocket_messages::{get_rabbitmq_uri, send_ws_message};
+use crate::messaging::{get_rabbitmq_uri, send_ws_message};
 
 #[allow(unused_must_use)]
 #[post(
