+++ rust/src/middleware/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod catchers;
+pub mod cors;
